PR Feedback (lucas): Handle case with MissingSidecar badging
To do this we needed to apply sidecar detection to outsider nodes,
meaning nodes outside of the requested graph namespaces (but still
in the accessible set of namespaces).

This led to a reworking of namespace-level appender caching for
workloads and service definitions.  The new approach better leverages
our underlying caching at the business layer, while still adding
some optimization if caching is disabled.

One more subtlety is that with this change we will treat
canonical_revision="latest" the same as app="unknown", because that
is how Istio is reporting and it is important when we do app workload
matching. But, this could affect a user that actually uses "latest"
as their actual version label value.